Fix vocab sizes wrt. reserved tokens
Different methods in LabelEncoder class were altering the final vocabulary size when self.max_size is set, because the reserved entries were not always handled identically, leading to more or fewer model parameters than requested depending on the options passed.

This commit attempts at fixing this by counting the reserved tokens as part of max_size.

In practice:
- compute_vocab will remove more entries to leave space for the reserved ones
- expand_vocab will substract the reserved tokens when counting the number of slots left
- expand_vocab will no longer erroneously expand the size of the vocab when max_size is set to shrink the new vocab size
- register_upper will no longer substract reserved entries twice to count the number of slots left

Additionally, in expand_vocab the min_freq condition is applied at the same time as the filtering to find new symbols, for optimization purpose