Implement full model finetuning (#9)
* Implement full model finetuning

Implemented via a new config parameter `load_pretrained_model`

Enable finetuning of another PaPie model by loading its state dict into the current model.

Customization of which parts to load is possible via the subparameter `load_pretrained_model["exclude"]`.

* SimpleModel load pretrained model refactor + improve tasks params loading

- change variable name `model_tar` to `pretrained`
- new nested function `load_state_dict_label_by_label` that applies to load wemb, cemb, lm and tasks with linear decoders label by label
- variable `model_parts_to_load` can only contain parts which are part of the model (e.g. "lm" only if `self.include_lm` is True)
- Load tasks of pretrained models even when the case don't match (e.g. "pos" task of the pretrained model can be loaded into the "POS" task of the new model)
- raise NotImplementedError if the task to be loaded does not correspond to a LinearDecoder (planned to implement AttentionalDecoder soon, otherwise the user should exclude the task in the config)

* Finetuning: implement load+expand pretrained label encoders

`MultiLabelEncoder.fit()` has a new boolean option `expand_mode` with which new labels/vocab are added to the freqs, then a new method `LabelEncoder.expand_vocab()` is called to extract a new list of labels/vocabs and append new entries at the end of `LabelEncoder.table` and `LabelEncoder.inverse_table` .

The `expand_mode` is optional and can be set to False/True (default True) in the json config ("load_pretrained_model"/"expand_labels"). If set to False in finetuning mode ("load_pretrained_model"/"pretrained":"pretrained_model.tar"), a new option `skip_fitted` is passed to `MultiLabelEncoder.fit()` instead, to be able to fit potential tasks that could be set in the new model but not in the pretrained one (example use case: fine-tune a model pretrained for POS tagging on a lemmatization task).

+ Fix typo in `LabelEncoder.from_json()`
+ reorganize imports in module dataset.py
+ fix logging.basicConfig by adding option force=True (fixes missing logs in the stdout)
+ change logs in module datasets.py with calls to a module-specific logger (best practice)

* LabelEncoder: Fix upper duplicates in register_upper()

Previously, if two vocabulary entries had the same uppercasing, the inverse_table would include the duplicates, but the table would include the upper entry only with the last index, leading to a missing index.

Found in an experiment where 2 characters were uppercased to "M", at positions 606 and 651, so table had only "M":651 with index 606 was missing, leading to random errors when the model tried to predict index 606.

* Finetune: replace option 'expand_labels' by 'labels_mode'

option 'labels_mode' takes 3 possible values:
- "expand" replaces expand_labels=true - append new vocab from the new data to the pretrained label encoders
- "skip" replaces expand_labels=false - only fit new tasks that haven't been pretrained
- "replace" fits a new MultiLabelEncoder (pretrained params will still be loaded for common vocab entries)

* Finetune: load tasks with AttentionalDecoder + fix missing sent RNN encoder loading

* Enable forcing seed from command-line or config file

User can pass a seed value either from the command-line with a new option `--seed` or from the config file, where the default value is 'auto'.

* Finetune: Improve logging when loading the pretrained parameters

* Finetune: Fix loading AttentionalDecoder tasks

* New utility script get_pie_model_params.py to show model info

This script can be useful in fine-tuning or reporting scenarios,
as it aims to show the size of the vocabularies as well as
the number and size of some important layers in the model.

* Finetune: fix reducing vocabulary

When "labels_mode" is "expand" but the vocabulary max sizes ("char_max_size"/"word_max_size") are smaller than the parent model's vocabularies, the intended behavior is to keep only the most frequent entries from the parent vocabulary.

A bug in the code led to removing the entire vocabulary, including the reserved entries (e.g. <UNK>). This commit fixes that.

* Add new finetune option 'replace_fill'

- New option 'replace_fill' enables replacing the vocab/labels with entries from the finetuning data, and then filling leftover spots with vocab/labels from the parent model
- Renamed variables in MultiLabelEncoder.fit() method + created a property that returns a list with all LabelEncoder objects stored in the MultiLabelEncoder
- Modify the `__main__` of train.py to reflect the behavior when launching training from group.py (for debugging)

* Rewrite LabelEncoder completion of mode 'replace_fill' using attribute freqs + refactor LabelEncoder.expand_vocab

Using the parent LabelEncoder.freqs attributes enables storing and using frequencies of vocab entries/labels. This way, the min_freq config option can be used, and only the most common parent items are added with the max_size option.

* train.py : Minor print message change

* Add logs during register_upper

To inform whether all available new uppercase entries could be registered or not.

* Improve logs during register_upper

* Fix vocab sizes wrt. reserved tokens

Different methods in LabelEncoder class were altering the final vocabulary size when self.max_size is set, because the reserved entries were not always handled identically, leading to more or fewer model parameters than requested depending on the options passed.

This commit attempts at fixing this by counting the reserved tokens as part of max_size.

In practice:
- compute_vocab will remove more entries to leave space for the reserved ones
- expand_vocab will substract the reserved tokens when counting the number of slots left
- expand_vocab will no longer erroneously expand the size of the vocab when max_size is set to shrink the new vocab size
- register_upper will no longer substract reserved entries twice to count the number of slots left

Additionally, in expand_vocab the min_freq condition is applied at the same time as the filtering to find new symbols, for optimization purpose

* register_upper - minor logs improvements

* Finetune: enable effectively excluding loading of parent weights in the wemb, cemb and task modules

* Update README - Finetuning section

* train.py: Rollback call to settings optimization pipeline

* README: Point to PaPie PyPI package